watch: Introduce `Watch::map_stream`
It's not unusual to want to lazily derive values from a `Watch`. For
example, when watching a resource (file, directory, etc) for updates, an
application may want to rebind services, etc.

This change introduces `Watch::map_stream` that converts a `Watch` into
a stream of values derived from borrows on the Watch.